pyproject.toml
[tool.poetry]
name = "pdb"
version = "0.1.0"
description = ""
authors = ["Ra <ravicity999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
biopython = "^1.84"
pandas = "^2.2.3"
pyarrow = "^18.0.0"
fastparquet = "^2024.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
